# api-specs/schack-se-api-docs.json
{
  "openapi": "3.0.1",
  "servers": [
    {
      "url": "https://member.schack.se/",
      "description": "Default Server URL"
    }
  ],
  "paths": {
...
    "/public/api/v1/ratinglist/federation/date/{ratingdate}/ratingtype/{ratingtype}/category/{category}": {
      "get": {
        "tags": [
          "rating-list-public-controller"
        ],
        "summary": "Hämta rankinglista för hela SSF, givet datum, rankingkategori och medlemskategori",
        "operationId": "getFederation",
        "parameters": [
          {
            "name": "ratingdate",
            "in": "path",
            "description": "datum för rankinglistan YYYY-MM-DD",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "ratingtype",
            "in": "path",
            "description": "rankingtyp, Långparti=1,Snabb=6,Blixt=7",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "category",
            "in": "path",
            "description": "Medlemskategori: Alla=0, Damer=5, Juniorer=1, Kadetter=2, Miniorer=6, Knattar=7, Veteraner=4",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PlayerInfoDto"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/public/api/v1/player/{id}/date/{date}": {
      "get": {
        "tags": [
          "player-public-controller"
        ],
        "summary": "Hämta spelarinfo",
        "operationId": "getPlayer",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id för spelare",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "date",
            "in": "path",
            "description": "datum för spelarinfo (klubbtillhörighet, ranking, etc) YYYY-MM-DD",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PlayerInfoDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/public/api/v1/player/fideid/{id}/date/{date}": {
      "get": {
        "tags": [
          "player-public-controller"
        ],
        "summary": "Hämta spelarinfo givet FIDE-id",
        "operationId": "getPlayerByFIDEId",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "FIDE-id för spelare",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "date",
            "in": "path",
            "description": "datum för spelarinfo (klubbtillhörighet, ranking, etc) YYYY-MM-DD",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PlayerInfoDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/public/api/v1/organisation/district/clubs/{districtid}": {
      "get": {
        "tags": [
          "organisation-public-controller"
        ],
        "summary": "Hämta klubbar i distrikt",
        "operationId": "getClubsInDistrict",
        "parameters": [
          {
            "name": "districtid",
            "in": "path",
            "description": "id för distriktet",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ClubDTO"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/public/api/v1/organisation/club/{clubid}": {
      "get": {
        "tags": [
          "organisation-public-controller"
        ],
        "summary": "Hämta klubb",
        "operationId": "getClub",
        "parameters": [
          {
            "name": "clubid",
            "in": "path",
            "description": "id för klubben",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ClubDTO"
                }
              }
            }
          }
        }
      }
    },
    "/public/api/v1/organisation/club/exists/{name}/{id}": {
      "get": {
        "tags": [
          "organisation-public-controller"
        ],
        "summary": "Kontrollera om klubbnamn finns, annat än för klubb-id't som skickas in",
        "operationId": "clubNameExists",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "description": "namn för klubben",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "in": "path",
            "description": "id för klubben som det frågas för",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "boolean"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "MemberFIDERatingDTO": {
        "type": "object",
        "properties": {
          "rating": {
            "type": "integer",
            "format": "int32"
          },
          "title": {
            "type": "string"
          },
          "date": {
            "type": "string",
            "format": "date-time"
          },
          "k": {
            "type": "integer",
            "format": "int32"
          },
          "rapidRating": {
            "type": "integer",
            "format": "int32"
          },
          "rapidk": {
            "type": "integer",
            "format": "int32"
          },
          "blitzRating": {
            "type": "integer",
            "format": "int32"
          },
          "blitzK": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "MemberLASKRatingDTO": {
        "type": "object",
        "properties": {
          "rating": {
            "type": "integer",
            "format": "int32"
          },
          "date": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "PlayerInfoDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "birthdate": {
            "type": "string",
            "format": "date-time"
          },
          "sex": {
            "type": "integer",
            "format": "int32"
          },
          "fideid": {
            "type": "integer",
            "format": "int32"
          },
          "country": {
            "type": "string"
          },
          "club": {
            "type": "string"
          },
          "clubId": {
            "type": "integer",
            "format": "int32"
          },
          "elo": {
            "$ref": "#/components/schemas/MemberFIDERatingDTO"
          },
          "lask": {
            "$ref": "#/components/schemas/MemberLASKRatingDTO"
          }
        }
      },
...
      "ClubDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string"
          },
          "street": {
            "type": "string"
          },
          "zipcode": {
            "type": "integer",
            "format": "int32"
          },
          "city": {
            "type": "string"
          },
          "startdate": {
            "type": "string",
            "format": "date-time"
          },
          "startSeason": {
            "type": "string",
            "format": "date-time"
          },
          "endSeason": {
            "type": "string",
            "format": "date-time"
          },
          "phonenr": {
            "type": "string"
          },
          "postgiro": {
            "type": "string"
          },
          "alliansclub": {
            "type": "integer",
            "format": "int32"
          },
          "email": {
            "type": "string"
          },
          "orgnumber": {
            "type": "string"
          },
          "districts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DistrictMembershipDTO"
            }
          },
          "regYear": {
            "$ref": "#/components/schemas/RegistrationYear"
          },
          "url": {
            "type": "string"
          },
          "vbdescr": {
            "type": "string"
          },
          "schoolClub": {
            "type": "integer",
            "format": "int32"
          },
          "schoolName": {
            "type": "string"
          },
          "schoolid": {
            "type": "integer",
            "format": "int32"
          },
          "lan": {
            "type": "integer",
            "format": "int32"
          },
          "noEconomy": {
            "type": "integer",
            "format": "int32"
          },
          "specialType": {
            "type": "integer",
            "format": "int32"
          },
          "yes2chessreason": {
            "type": "boolean"
          },
          "schack4anreason": {
            "type": "boolean"
          },
          "ovrigtreason": {
            "type": "boolean"
          },
          "hasRatingPlayers": {
            "type": "integer",
            "format": "int32"
          },
          "sisuid": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "DistrictMembershipDTO": {
        "type": "object",
        "properties": {
          "start": {
            "type": "string",
            "format": "date-time"
          },
          "end": {
            "type": "string",
            "format": "date-time"
          },
          "year": {
            "type": "integer",
            "format": "int32"
          },
          "districtid": {
            "type": "integer",
            "format": "int32"
          },
          "clubid": {
            "type": "integer",
            "format": "int32"
          },
          "active": {
            "type": "integer",
            "format": "int32"
          },
          "benefit": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "RegistrationYear": {
        "type": "object",
        "properties": {
          "startDate": {
            "type": "string",
            "format": "date-time"
          },
          "endDate": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    }
  }
}